Send x,y player data to server
Cannot host more than one lobby. This is because the “hostnumber” is set wrong and I have zero idea why so I'm setting it as 0 for now so I can turn something in. Other than this the server is taking the correct data and the client is sending the correct data. (should be an easy fix, I just do not have the time nor the energy at this time)